Update README.md
This Power BI report presents the results of a sentiment analysis conducted on tweets from a Hindi newspaper. The purpose of the analysis is to examine the sentiments expressed by Twitter users regarding different topics covered by the newspaper. The report offers a detailed analysis of sentiment trends, distribution, and drivers within the tweets. Moreover, it includes slicers, which enable users to filter the graphs based on weekdays and sentiment categories, facilitating a more targeted exploration of the data.

The report begins by providing an overview of the sentiment analysis, explaining the methodology used to classify tweets as positive, negative, or neutral. It also highlights the importance of sentiment analysis in understanding public opinion and shaping decision-making processes.

The sentiment trends section of the report showcases a line graph that illustrates the overall sentiment score over time. This graph allows users to identify any notable shifts in sentiment towards the newspaper's topics. By utilizing the slicers, users can focus on specific weekdays or sentiment categories, enabling a more granular analysis of sentiment patterns.

The sentiment distribution section presents a pie chart that visualizes the distribution of positive, negative, and neutral sentiments within the tweets. This chart provides a quick snapshot of the overall sentiment distribution, allowing users to assess the general sentiment landscape at a glance. Again, slicers can be utilized to narrow down the analysis based on specific weekdays or sentiment categories.

The sentiment drivers section of the report delves deeper into the factors influencing sentiment within the tweets. It utilizes stacked bar charts to showcase the top keywords associated with positive and negative sentiments. By examining these keywords, users can gain insights into the specific topics or events driving positive or negative sentiment.

Overall, this Power BI report offers a comprehensive analysis of sentiment expressed in tweets related to a Hindi newspaper. It empowers users to explore sentiment trends, distribution, and drivers in a flexible and interactive manner. The inclusion of slicers enhances the report's usability, enabling users to filter and drill down into specific aspects of the data. With its user-friendly interface and insightful visualizations, this report provides valuable insights into public sentiment towards various topics covered by the Hindi newspaper.